src/requirements-docker.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.2
build==1.0.3
CacheControl==0.13.1
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.6
chromadb==0.5.7
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
crashtest==0.4.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
dulwich==0.21.7
durationpy==0.7
executing==2.0.0
fastapi==0.115.0
fastjsonschema==2.18.1
filelock==3.13.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.0.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.4
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==24.3.0
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.0
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.2
langchain-text-splitters==0.3.0
langsmith==0.1.125
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.22.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
oauthlib==3.2.2
onnxruntime==1.19.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.5.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
posthog==3.6.6
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.5
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyfiglet==1.0.2
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.1.1
pypdf==5.0.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rapidfuzz==3.5.2
referencing==0.30.2
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.10.6
rsa==4.9
Send2Trash==1.8.2
setuptools==75.1.0
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
stack-data==0.6.3
starlette==0.38.5
sympy==1.13.3
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.20.0
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
trove-classifiers==2023.11.29
typer==0.12.5
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.30.6
virtualenv==20.24.7
watchfiles==0.24.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==13.0.1
Werkzeug==3.0.4
widgetsnbextension==4.0.9
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.2
